feat: improve pr body output
refactor: clean up postLangPRs, rm no-longer-used util functions, add decompress package types
