+++ package.json
@@ -66,6 +66,7 @@
     "@storybook/react": "7.6.6",
     "@storybook/testing-library": "0.2.2",
     "@svgr/webpack": "^8.1.0",
+    "@types/decompress": "^4.2.7",
     "@types/hast": "^3.0.0",
     "@types/node": "^20.4.2",
     "@types/react": "^18.2.15",
